getParkingFee method updated

+++ src/main/java/service/ExitService.java
@@ -3,5 +3,5 @@ package service;
 import model.ParkingTicket;
 
 public interface ExitService extends UpdateParkingSpot {
-    Long getParkingFee(ParkingTicket parkingTicket);
+    Double getParkingFee(ParkingTicket parkingTicket);
 }
